requirements.txt
arcade==3.3.1
pyglet==2.1.5
pytiled-parser==2.2.9
python-dotenv
groq